refactor: move default before to file / removed unsed files
