--- src/app.controller.ts
@@ -1,9 +1,0 @@
-import { Controller, Get } from '@nestjs/common';
-
-@Controller()
-export class AppController {
-  @Get()
-  getHello() {
-    return 'Hello World!';
-  }
-}
