--- test/app.e2e-spec.ts
@@ -1,25 +1,0 @@
-import * as request from 'supertest';
-import { Test } from '@nestjs/testing';
-import { INestApplication } from '@nestjs/common';
-import * as redis from 'redis-mock';
-import { AppModule } from '../src/app.module';
-
-describe('AppController (e2e)', () => {
-  let app: INestApplication;
-
-  beforeAll(async () => {
-    await redis.createClient();
-
-    const moduleFixture = await Test.createTestingModule({
-      imports: [AppModule],
-    }).compile();
-
-    app = moduleFixture.createNestApplication();
-    await app.init();
-  });
-
-  it('/ (GET)', () => request(app.getHttpServer())
-    .get('/')
-    .expect(200)
-    .expect('Hello World!'));
-});
